feat: save bot email in ssm param
