+++ justfile
@@ -542,9 +542,17 @@ auth-import ci_dir="sandbox-ci":
 auth-check:
     uv run python scripts/sync_auth_state.py check
 
+# Print the GitHub bot account password from CI secrets
+auth-bot-password ci_dir="sandbox-ci":
+    @uv run python scripts/auth_bot_secret.py {{ci_dir}} password
+
 # Generate a 2FA code for the GitHub bot account (requires oathtool)
-auth-2fa-code ci_dir="sandbox-ci":
-    @uv run python scripts/auth_2fa_code.py {{ci_dir}}
+auth-bot-2fa ci_dir="sandbox-ci":
+    @uv run python scripts/auth_bot_secret.py {{ci_dir}} totp
+
+# Print the GitHub bot account email from CI project variables
+auth-bot-email ci_dir="sandbox-ci":
+    @uv run jd show -v github_bot_account_email --text --path {{ci_dir}}
 
 # Generate .env for base template E2E tests from deployed project + CI infrastructure
 # Reads variables from the project, OAuth creds from CI, and accepts user options
