+++ libs/jupyter-infra-tf-aws-iam-ci/jupyter_infra_tf_aws_iam_ci/template/engine/outputs.tf
@@ -37,6 +37,12 @@ output "github_bot_account_totp_secret_secret_arn" {
   value       = module.github_bot_account_totp_secret.secret_arn
 }
 
+# GitHub bot account email — SSM parameter
+output "github_bot_account_email_arn" {
+  description = "ARN of the SSM parameter for GitHub bot account email."
+  value       = module.github_bot_account_email.parameter_arn
+}
+
 # OAuth app client IDs (x5) — SSM parameters
 output "github_oauth_app_client_id_1_arn" {
   description = "ARN of the SSM parameter for GitHub OAuth app #1 client ID."
